requirements.txt
altair                    ==5.5.0
attrs                     ==25.1.0
blinker                   ==1.9.0
cachetools                ==5.5.2
certifi                   ==2025.1.31
charset-normalizer        ==3.4.1
click                     ==8.1.8
filelock                  ==3.17.0
fsspec                    ==2025.2.0
fugashi                   ==1.4.0
gitdb                     ==4.0.12
GitPython                 ==3.1.44
huggingface-hub           ==0.29.2
idna                      ==3.10
Jinja2                    ==3.1.6
jsonschema                ==4.23.0
jsonschema-specifications ==2024.10.1
MarkupSafe                ==3.0.2
mpmath                    ==1.3.0
narwhals                  ==1.29.0
networkx                  ==3.2.1
numpy                     ==1.26.4
packaging                 ==24.2
pandas                    ==2.2.3
pillow                    ==11.1.0
pinecone                  ==6.0.1
pinecone-plugin-interface ==0.0.7
pip                       ==25.0.1
plotly                    ==6.0.0
protobuf                  ==5.29.3
pyarrow                   ==19.0.1
pydeck                    ==0.9.1
python-dateutil           ==2.9.0.post0
python-dotenv             ==1.0.1
pytz                      ==2025.1
PyYAML                    ==6.0.2
referencing               ==0.36.2
regex                     ==2024.11.6
requests                  ==2.32.3
rpds-py                   ==0.23.1
safetensors               ==0.5.3
sentencepiece             ==0.2.0
setuptools                ==58.0.4
six                       ==1.17.0
smmap                     ==5.0.2
st-cytoscape              ==0.0.5
streamlit                 ==1.43.0
sympy                     ==1.13.1
tenacity                  ==9.0.0
tokenizers                ==0.21.0
toml                      ==0.10.2
torch                     ==2.6.0
torchaudio                ==2.6.0
torchvision               ==0.21.0
tornado                   ==6.4.2
tqdm                      ==4.67.1
transformers              ==4.49.0
typing_extensions         ==4.12.2
tzdata                    ==2025.1
unidic-lite               ==1.0.8
urllib3                   ==2.3.0